+++ src/app/question/hooks/useRecordForm.ts
@@ -1,0 +1,34 @@
+import { useState } from 'react';
+import { useTimer } from './useTimer';
+
+export const INPUT_MAX_LENGTH = 100;
+
+export const useRecordForm = () => {
+  const { setIsActive, remainingSeconds, scheduleSetInactive, isTimerEnd } =
+    useTimer();
+
+  const [input, setInput] = useState<string>('');
+
+  const handleInputFocus = () => {
+    setIsActive(true);
+  };
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+    setInput(e.target.value);
+    setIsActive(true);
+    scheduleSetInactive();
+  };
+
+  const handleInputBlur = () => {
+    setIsActive(false);
+  };
+
+  return {
+    remainingSeconds,
+    isTimerEnd,
+    input,
+    handleInputFocus,
+    handleInputChange,
+    handleInputBlur,
+  };
+};
