+++ src/app/question/hooks/useTimer.ts
@@ -2,14 +2,12 @@ import { useEffect, useRef, useState } from 'react';
 
 const INITIAL_REMAINING_SECONDS = 10;
 const INACTIVE_THRESHOLD_SECONDS = 1;
-export const INPUT_MAX_LENGTH = 100;
 
-export const useQuestionForm = () => {
+export const useTimer = () => {
   const [isActive, setIsActive] = useState<boolean>(false);
   const [remainingSeconds, setRemainingSeconds] = useState<number>(
     INITIAL_REMAINING_SECONDS,
   );
-  const [input, setInput] = useState<string>('');
 
   const timerRef = useRef<NodeJS.Timeout | null>(null);
   const debounceRef = useRef<NodeJS.Timeout | null>(null);
@@ -43,20 +41,6 @@ export const useQuestionForm = () => {
     }, INACTIVE_THRESHOLD_SECONDS * 1000);
   };
 
-  const handleInputFocus = () => {
-    setIsActive(true);
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setInput(e.target.value);
-    setIsActive(true);
-    scheduleSetInactive();
-  };
-
-  const handleInputBlur = () => {
-    setIsActive(false);
-  };
-
   useEffect(() => {
     if (isActive) {
       clearTimer();
@@ -82,11 +66,9 @@ export const useQuestionForm = () => {
   }, []);
 
   return {
+    setIsActive,
     remainingSeconds,
-    input,
+    scheduleSetInactive,
     isTimerEnd,
-    handleInputFocus,
-    handleInputChange,
-    handleInputBlur,
   };
 };
